composition_engine/requirements.txt
boto3==1.34.34
langchain==0.1.9
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.1 